requirements.txt
ffmpeg-python==0.2.0
demucs==4.0.0
pydub==0.25.1
moviepy==1.0.3
whisperx>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0
groq>=0.4.2
indic-transliteration>=2.3.56
TTS==0.22.0
tqdm==4.66.1
pyyaml==6.0.1
gradio==4.19.2
opencv-python==4.8.1.78
scipy==1.11.4
face_alignment==1.3.5
imageio==2.33.1
imageio-ffmpeg==0.4.9
librosa==0.10.1
